# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Korean Learning API",
    "description": "API REST Gin/SQLite pour apprendre le coreen avec decks, flashcards, revision espacee et authentification JWT.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Serveur local go run ."
    }
  ],
...
  "paths": {
...
    "/user/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Creer un utilisateur",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/AuthResult" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Connecter un utilisateur",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/AuthResult" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/user/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Lire son profil",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Profil public",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PublicUser" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "put": {
        "tags": ["Auth"],
        "summary": "Mettre a jour son profil",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateUserRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profil mis a jour",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PublicUser" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/admin/users/{id}": {
      "put": {
        "tags": ["Admin"],
        "summary": "Modifier un utilisateur non admin",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateUserRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Utilisateur modifie",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PublicUser" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/admin/reset": {
      "post": {
        "tags": ["Admin"],
        "summary": "Reinitialiser la base de donnees depuis l'espace admin",
        "description": "Supprime decks, cartes, reviews et utilisateurs non admin. Les administrateurs sont conserves.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/ResetResult" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "delete": {
        "tags": ["Admin"],
        "summary": "Reinitialiser la base de donnees depuis l'espace admin",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/ResetResult" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/reset": {
      "post": {
        "tags": ["Admin"],
        "summary": "Reinitialiser la base de donnees",
        "description": "Supprime decks, cartes, reviews et utilisateurs non admin. Les administrateurs sont conserves.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/ResetResult" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "delete": {
        "tags": ["Admin"],
        "summary": "Reinitialiser la base de donnees",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/ResetResult" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/decks": {
      "get": {
        "tags": ["Decks"],
        "summary": "Lister tous les decks",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Liste des decks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Deck" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Decks"],
        "summary": "Creer un deck",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DeckRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Deck cree",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Deck" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/decks/search": {
      "get": {
        "tags": ["Decks", "Search"],
        "summary": "Rechercher des decks sur tous leurs champs",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Query" }],
        "responses": {
          "200": {
            "description": "Decks trouves",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Deck" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/decks/{id}": {
      "get": {
        "tags": ["Decks"],
        "summary": "Afficher un deck par ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "responses": {
          "200": {
            "description": "Deck",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Deck" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["Decks"],
        "summary": "Mettre a jour un deck",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DeckPatchRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deck modifie",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Deck" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Decks"],
        "summary": "Supprimer un deck",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "responses": {
          "204": { "description": "Deck supprime" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/decks/bulk": {
      "put": {
        "tags": ["Decks"],
        "summary": "Mettre a jour plusieurs decks par ID",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BulkDeckUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Decks modifies",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Deck" }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      },
      "delete": {
        "tags": ["Decks"],
        "summary": "Supprimer plusieurs decks par ID",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BulkDeleteRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/DeleteCount" },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/cards": {
      "get": {
        "tags": ["Cards"],
        "summary": "Lister toutes les cartes",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Liste des cartes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Card" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Cards"],
        "summary": "Creer une carte",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CardRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Carte creee",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Card" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/cards/search": {
      "get": {
        "tags": ["Cards", "Search"],
        "summary": "Rechercher des cartes sur tous leurs champs",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Query" }],
        "responses": {
          "200": {
            "description": "Cartes trouvees",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Card" }
                }
              }
            }
          }
        }
      }
    },
    "/api/cards/difficult": {
      "get": {
        "tags": ["Cards", "Stats"],
        "summary": "Lister les cartes difficiles",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Cartes avec au moins deux echecs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Card" }
                }
              }
            }
          }
        }
      }
    },
    "/api/cards/{id}": {
      "get": {
        "tags": ["Cards"],
        "summary": "Afficher une carte par ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "responses": {
          "200": {
            "description": "Carte",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Card" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["Cards"],
        "summary": "Mettre a jour une carte",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CardPatchRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Carte modifiee",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Card" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Cards"],
        "summary": "Supprimer une carte",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "responses": {
          "204": { "description": "Carte supprimee" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/api/search": {
      "get": {
        "tags": ["Search"],
        "summary": "Recherche globale decks et cartes",
        "description": "Recherche la sous-chaine dans les champs des decks et des cartes.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Query" }],
        "responses": {
          "200": {
            "description": "Resultats groupes par ressource",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SearchResult" }
              }
            }
          }
        }
      }
    },
    "/search": {
      "get": {
        "tags": ["Search"],
        "summary": "Alias de recherche globale",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Query" }],
        "responses": {
          "200": {
            "description": "Resultats groupes par ressource",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SearchResult" }
              }
            }
          }
        }
      }
    },
...
    "/api/stats": {
      "get": {
        "tags": ["Stats"],
        "summary": "Afficher les statistiques",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Statistiques",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StudyStats" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "responses": {
      "AuthResult": {
        "description": "JWT et utilisateur public",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/AuthResult" }
          }
        }
      },
...
      "BadRequest": {
        "description": "Requete invalide",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
      "Conflict": {
        "description": "Conflit de donnees",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "schemas": {
...
      "RegisterRequest": {
        "type": "object",
        "required": ["name", "email", "password"],
        "properties": {
          "name": { "type": "string", "minLength": 2, "maxLength": 80, "example": "Arthur" },
          "email": { "type": "string", "format": "email", "example": "arthur@example.com" },
          "password": { "type": "string", "format": "password", "minLength": 8, "example": "password123" }
        }
      },
...
    }
  }
}